Add check user is valid

+++ src/main/java/com/sonphattran/resumeportal/controllers/UpdateController.java
@@ -1,11 +1,14 @@
 package com.sonphattran.resumeportal.controllers;
 
+import com.sonphattran.resumeportal.authentication.AuthenticationFacade;
 import com.sonphattran.resumeportal.models.Education;
 import com.sonphattran.resumeportal.models.Experience;
 import com.sonphattran.resumeportal.models.Skill;
 import com.sonphattran.resumeportal.models.User;
 import com.sonphattran.resumeportal.services.user.*;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.view.RedirectView;
@@ -15,6 +18,9 @@ import java.util.ArrayList;
 
 @Controller
 public class UpdateController {
+    @Autowired
+    private AuthenticationFacade authenticationFacade;
+
     @Autowired
     private IUserService userService;
 
@@ -27,8 +33,35 @@ public class UpdateController {
     @Autowired
     private IExperienceService experienceService;
 
+    // Check if the current logged-in user has the right to modify the content
+    private boolean isLoggedIn() {
+        Authentication authentication = authenticationFacade.getAuthentication();
+        return authentication != null && !(authentication instanceof AnonymousAuthenticationToken);
+    }
+
+    private User getLoggedInUser() {
+        Authentication authentication = authenticationFacade.getAuthentication();
+        String username = (String) authentication.getPrincipal();
+        return userService.findUserByName(username);
+    }
+
+    private boolean isValidUser(Long userId) {
+        if (!isLoggedIn()) {
+            return false;
+        }
+
+        // Get the current logged-in user and check
+        User loggedInUser = getLoggedInUser();
+        return loggedInUser.getId() == userId;
+    }
+
     @PostMapping("/users/{userId}")
     public RedirectView update(@PathVariable Long userId, @ModelAttribute("user") User user) {
+        // Check if is valid
+        if (!isValidUser(userId)) {
+            return new RedirectView("home?invalid");
+        }
+
         // Update
         userService.updateName(
                 userId,
@@ -48,6 +81,11 @@ public class UpdateController {
             @RequestParam(value = "new_period", required = false) String[] period,
             @RequestParam(value = "new_details", required = false) String[] details
     ) {
+        // Check if is valid
+        if (!isValidUser(userId)) {
+            return new RedirectView("home?invalid");
+        }
+
         // Set user
         user.setId(userId);
 
@@ -83,6 +121,11 @@ public class UpdateController {
 
     @GetMapping("/users/{userId}/education/delete/{educationId}")
     public RedirectView removeEducation(@PathVariable Long userId, @PathVariable Long educationId) {
+        // Check if is valid
+        if (!isValidUser(userId)) {
+            return new RedirectView("home?invalid");
+        }
+
         // Remove education id
         educationService.deleteEducationById(educationId);
         return new RedirectView("/home#educations");
@@ -93,6 +136,11 @@ public class UpdateController {
             @PathVariable Long userId,
             @ModelAttribute("user") User user,
             @RequestParam(value = "new_skill", required = false) String[] newSkill) {
+        // Check if is valid
+        if (!isValidUser(userId)) {
+            return new RedirectView("home");
+        }
+
         // Set user
         user.setId(userId);
 
@@ -129,6 +177,11 @@ public class UpdateController {
     public RedirectView removeSkill(
             @PathVariable Long userId,
             @PathVariable Long skillId) {
+        // Check if is valid
+        if (!isValidUser(userId)) {
+            return new RedirectView("home?invalid");
+        }
+
         skillService.deleteSkillById(skillId);
         return new RedirectView("/home#skills");
     }
@@ -142,6 +195,11 @@ public class UpdateController {
             @RequestParam(value = "new_exp_period", required = false) String[] periods,
             @RequestParam(value = "new_desc", required = false) String[] descriptions
             ) {
+        // Check if is valid
+        if (!isValidUser(userId)) {
+            return new RedirectView("home?invalid");
+        }
+
         // Set user id
         user.setId(userId);
 
@@ -178,12 +236,22 @@ public class UpdateController {
     public RedirectView removeExperience(
             @PathVariable Long userId,
             @PathVariable Long experienceId) {
+        // Check if is valid
+        if (!isValidUser(userId)) {
+            return new RedirectView("home?invalid");
+        }
+
         experienceService.deleteExperienceById(experienceId);
         return new RedirectView("/home#experiences");
     }
 
     @PostMapping("/users/{userId}/visibility")
     public RedirectView updateVisibility(@PathVariable Long userId, @ModelAttribute("user") User user) {
+        // Check if is valid
+        if (!isValidUser(userId)) {
+            return new RedirectView("home?invalid");
+        }
+
         userService.updateVisibility(userId, user.isVisible());
         return new RedirectView("/home#visibility-section");
     }
